Move ProfilingExperiment and ProfilingLog to their own file.
